Refactor scheduler and console management functionality
- Changed `findProcess` visibility in `AScheduler.h` to public.
- Updated `createBaseScreen` in `ConsoleManager` to accept a new parameter for console switching.
- Commented out `execute` method and constructor in `DeclareCommand` files for temporary disablement.
- Implemented First-Come, First-Served scheduling in `FCFSScheduler` with new methods for process management.
- Modified `GlobalScheduler` to initialize and manage the lifecycle of the `FCFSScheduler`.
- Updated command handling in `MainConsole` to support new scheduler commands.
- Refined `executeCurrentCommand` in `Process` for better command execution control.
- Enhanced `SchedulerWorker` to support threading for process execution.
- Adjusted `SleepCommand` constructor to use `uint8_t` for `sleepTicks` to optimize memory usage.

+++ SleepCommand.h
@@ -3,7 +3,7 @@
 class SleepCommand : public ICommand
 {
 public:
-	SleepCommand(int processID, uint16_t sleepTicks) : ICommand(processID, CommandType::SLEEP) 
+	SleepCommand(int processID, uint8_t sleepTicks) : ICommand(processID, CommandType::SLEEP) 
 	{
 		this->processID = processID;
 		this->sleepTicks = sleepTicks;
